changed cardNumber datatype to LONG from INT

+++ geekText/src/main/java/org/geektext/model/CreditCard.java
@@ -4,14 +4,14 @@ import jakarta.persistence.OneToOne;
 
 public class CreditCard {
 
-    private final int cardNumber;
+    private final long cardNumber;
     private final int cvv;
     private final int expDate;
 
     @OneToOne
     private User user;
 
-    public CreditCard(int cardNumber, int cvv, int expDate, User user) {
+    public CreditCard(long cardNumber, int cvv, int expDate, User user) {
         this.user = user;
         this.cardNumber = cardNumber;
         this.cvv = cvv;
@@ -25,7 +25,7 @@ public class CreditCard {
         this.user = user;
     }
 
-    public int getCardNumber() {
+    public long getCardNumber() {
         return cardNumber;
     }
